// src/app/interfaces/interfacesUsuarios.ts
export interface Usuarios {
    Id_Usuario: string;
    Identificador: string;
    Nombres: string;
    Apellidos: string;
    Cedula: string;
    Correo_Electronico: string;
    Fecha_Nacimiento: string;
    Domicilio: string;
    Celular: string;
    Estado_Vacunacion: boolean;
    Tipo_Vacuna: string;
    Fecha_Vacunacion: string;
    Numero_Dosis: string;
    Usuario: string;
    Password: string;
}